refactor: PasswordResetTokenEntity

+++ src/main/java/br/com/juliocauan/authentication/domain/model/PasswordResetToken.java
@@ -7,10 +7,11 @@ public abstract class PasswordResetToken {
     protected static final int TOKEN_LENGTH = 32;
     public static final int TOKEN_EXPIRATION_MINUTES = 10;
 
-    public abstract Long getId();
+    public abstract Integer getId();
     public abstract User getUser();
     public abstract String getToken();
     public abstract LocalDateTime getExpireDate();
+    
     public final boolean isExpired() {
         return LocalDateTime.now().isAfter(getExpireDate());
     }
