+++ src/main/java/br/com/juliocauan/authentication/infrastructure/repository/PasswordResetTokenRepositoryImpl.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.juliocauan.authentication.domain.repository.PasswordResetTokenRepository;
 import br.com.juliocauan.authentication.infrastructure.model.PasswordResetTokenEntity;
 
-public interface PasswordResetTokenRepositoryImpl extends PasswordResetTokenRepository, JpaRepository<PasswordResetTokenEntity, Long> {
+public interface PasswordResetTokenRepositoryImpl extends PasswordResetTokenRepository, JpaRepository<PasswordResetTokenEntity, Integer> {
     
 }
